# architecture/api-contract-openapi.yaml
openapi: 3.0.3
info:
  title: iReg Moto BF API
  description: >
    API réglementaire pour la conformité des deux-roues motorisés au Burkina Faso.
    Plateforme SaaS interfaçant importateurs, distributeurs, unités d'assemblage,
    détaillants et le ministère des transports (DRCTT).
    Supporte REST (mutations métier) et GraphQL (analytics).
  version: 1.0.0
  contact:
    name: iReg Moto BF Technical Team
    email: tech@ireg-moto.bf
  license:
    name: Proprietary
    url: https://ireg-moto.bf/licence
servers:
  - url: https://api.ireg-moto.bf/v1
    description: Production
  - url: https://api-staging.ireg-moto.bf/v1
    description: Staging
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /vehicles:
    get:
      tags: [Vehicules]
      summary: Liste des véhicules
      description: Recherche et liste des véhicules avec filtres avancés.
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
        - name: vin
          in: query
          schema: { type: string, pattern: '^[A-HJ-NPR-Z0-9]{17}$' }
          description: Recherche par VIN exact
        - name: status
          in: query
          schema:
            type: string
            enum: [IMPORTED, IN_STOCK, RESERVED, SOLD, REGISTERED, STOLEN, SCRAPPED, EXPORTED]
        - name: category
          in: query
          schema:
            type: string
            enum: [CYCLOMOTEUR_50CC, MOTO_LEGERE_125CC, MOTO_LOURDE_125CC_PLUS, TRICYCLE_MOTEUR, QUADRACYCLE, ELECTRIQUE_50CC_EQ, ELECTRIQUE_125CC_EQ, ELECTRIQUE_PLUS]
        - name: manufacturer
          in: query
          schema: { type: string }
        - name: model
          in: query
          schema: { type: string }
        - name: actorId
          in: query
          schema: { type: string, format: uuid }
          description: Filtrer par acteur propriétaire
        - name: warehouseId
          in: query
          schema: { type: string, format: uuid }
        - name: importDateFrom
          in: query
          schema: { type: string, format: date }
        - name: importDateTo
          in: query
          schema: { type: string, format: date }
        - name: isBlacklisted
          in: query
          schema: { type: boolean }
      responses:
        '200':
          description: Liste des véhicules
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedVehicles'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Vehicules]
      summary: Enregistrer un véhicule (import/assemblage)
      description: >
        Enregistre un nouveau véhicule dans le système.
        Vérifie automatiquement que la catégorie est autorisée
        et que l'homologation est valide.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVehicleRequest'
      responses:
        '201':
          description: Véhicule enregistré
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          description: Véhicule non conforme (catégorie interdite)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /vehicles/{vin}:
    parameters:
      - name: vin
        in: path
        required: true
        schema:
          type: string
          pattern: '^[A-HJ-NPR-Z0-9]{17}$'

    get:
      tags: [Vehicules]
      summary: Détail d'un véhicule par VIN
      description: >
        Retourne les informations complètes d'un véhicule
        incluant son historique de propriété et son statut de conformité.
      responses:
        '200':
          description: Détail véhicule
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Vehicules]
      summary: Modifier un véhicule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVehicleRequest'
      responses:
        '200':
          description: Véhicule mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /vehicles/{vin}/qr-generate:
    parameters:
      - name: vin
        in: path
        required: true
        schema:
          type: string
          pattern: '^[A-HJ-NPR-Z0-9]{17}$'
    post:
      tags: [Vehicules]
      summary: Générer un QR code pour un véhicule
      description: >
        Génère un QR code unique contenant les informations
        essentielles du véhicule (VIN, catégorie, conformité).
      responses:
        '200':
          description: QR code généré
          content:
            application/json:
              schema:
                type: object
                properties:
                  qrCode:
                    type: string
                    description: Données encodées dans le QR
                  qrImageUrl:
                    type: string
                    format: uri
                    description: URL de l'image PNG du QR code
                  generatedAt:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /vehicles/blacklist:
    get:
      tags: [Vehicules]
      summary: Liste noire des véhicules
      description: Liste des véhicules signalés (vol, fraude, contrefaçon).
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: reason
          in: query
          schema:
            type: string
            enum: [STOLEN, FRAUD, COUNTERFEIT, NON_COMPLIANT, ADMIN_SEIZURE, INSURANCE_TOTAL_LOSS]
        - name: search
          in: query
          schema: { type: string }
          description: Recherche par VIN
        - name: isActive
          in: query
          schema: { type: boolean, default: true }
      responses:
        '200':
          description: Liste noire
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/VehicleBlacklistEntry'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Vehicules]
      summary: Ajouter un véhicule à la liste noire
      description: >
        Signale un véhicule comme volé, frauduleux ou contrefait.
        Nécessite le rôle CONTROLEUR, CNTI_AGENT ou ADMIN.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBlacklistEntryRequest'
      responses:
        '201':
          description: Véhicule ajouté à la liste noire
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleBlacklistEntry'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ==========================================================================
  # STOCKS (MODULE B)
  # ==========================================================================
  /stocks/movements:
    get:
      tags: [Stocks]
      summary: Mouvements de stock
      description: Historique des mouvements de stock avec filtres.
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: vehicleId
          in: query
          schema: { type: string, format: uuid }
        - name: actorId
          in: query
          schema: { type: string, format: uuid }
        - name: warehouseId
          in: query
          schema: { type: string, format: uuid }
        - name: movementType
          in: query
          schema:
            type: string
            enum: [IMPORT, ASSEMBLY, TRANSFER, SALE, RETURN, ADJUSTMENT, SCRAP, RESERVATION, RELEASE]
        - name: dateFrom
          in: query
          schema: { type: string, format: date }
        - name: dateTo
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: Liste des mouvements
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/StockMovement'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Stocks]
      summary: Enregistrer un mouvement de stock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStockMovementRequest'
      responses:
        '201':
          description: Mouvement enregistré
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockMovement'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          description: Stock insuffisant ou véhicule non disponible
  /stocks/inventory:
    get:
      tags: [Stocks]
      summary: Inventaires de stock
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: warehouseId
          in: query
          schema: { type: string, format: uuid }
        - name: status
          in: query
          schema:
            type: string
            enum: [PLANNED, IN_PROGRESS, COMPLETED, RECONCILED, DISCREPANCY_FOUND]
      responses:
        '200':
          description: Liste des inventaires
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/InventoryCount'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Stocks]
      summary: Créer un inventaire
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [warehouseId, plannedDate]
              properties:
                warehouseId:
                  type: string
                  format: uuid
                plannedDate:
                  type: string
                  format: date
                notes:
                  type: string
      responses:
        '201':
          description: Inventaire créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryCount'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /stocks/summary:
    get:
      tags: [Stocks]
      summary: Résumé des stocks par entrepôt
      description: >
        Retourne un résumé des stocks par catégorie et entrepôt.
        Utilisé pour le tableau de bord.
      parameters:
        - name: actorId
          in: query
          required: true
          schema: { type: string, format: uuid }
        - name: warehouseId
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Résumé des stocks
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        category:
                          type: string
                        categoryLabel:
                          type: string
                        totalUnits:
                          type: integer
                        totalValue:
                          type: number
                        warehouseId:
                          type: string
                        warehouseName:
                          type: string
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ==========================================================================
  # CLIENTS (MODULE C)
  # ==========================================================================
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT d'accès (RS256)
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Clé API pour les intégrations système (ministère, DGI)

  # ========================================================================
  # RESPONSES
  # ========================================================================
  responses:
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Non authentifié — Token invalide ou expiré
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Accès interdit — Permissions insuffisantes
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Ressource non trouvée
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Error:
      type: object
      required: [statusCode, message]
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: "Le VIN fourni n'est pas valide"
        code:
          type: string
          example: INVALID_VIN
        details:
          type: object
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          example: /api/v1/vehicles
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean

    # -------------------------------------------------------------------------
    # Auth
    # -------------------------------------------------------------------------
...
    Vehicle:
      type: object
      properties:
        id:
          type: string
        vin:
          type: string
        chassisNumber: { type: string }
        engineNumber: { type: string }
        manufacturer: { type: string }
        model: { type: string }
        modelYear: { type: integer }
        category: { type: string }
        categoryLabel: { type: string }
        fuelType: { type: string }
        transmission: { type: string }
        cylinderCapacityCc: { type: integer }
        powerKw: { type: number }
        weightKg: { type: number }
        color: { type: string }
        status: { type: string }
        qrCode: { type: string }
        homologationNumber: { type: string }
        homologationValidUntil:
          type: string
          format: date
        complianceStatus: { type: string }
        importDate:
          type: string
          format: date
        customsValue: { type: number }
        currentOwnerActorId: { type: string }
        currentWarehouseId: { type: string }
        createdAt:
          type: string
          format: date-time
    VehicleDetail:
      allOf:
        - $ref: '#/components/schemas/Vehicle'
        - type: object
          properties:
            ownershipHistory:
              type: array
              items:
                $ref: '#/components/schemas/VehicleOwnership'
            isBlacklisted:
              type: boolean
            blacklistInfo:
              type: object
              properties:
                reason: { type: string }
                reasonDetails: { type: string }
                reportedAt:
                  type: string
                  format: date-time
    PaginatedVehicles:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Vehicle'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    CreateVehicleRequest:
      type: object
      required: [vin, manufacturer, model, modelYear, categoryId, fuelType, cylinderCapacityCc]
      properties:
        vin:
          type: string
          pattern: '^[A-HJ-NPR-Z0-9]{17}$'
        chassisNumber: { type: string }
        engineNumber: { type: string }
        manufacturer: { type: string }
        model: { type: string }
        modelYear:
          type: integer
          minimum: 1990
          maximum: 2030
        categoryId:
          type: string
          format: uuid
        fuelType:
          type: string
          enum: [ESSENCE, DIESEL, ELECTRIQUE, HYBRIDE, GPL]
        transmission:
          type: string
          enum: [MANUELLE, AUTOMATIQUE, SEMI_AUTOMATIQUE, CONTINUOUSLY_VARIABLE]
        cylinderCapacityCc: { type: integer }
        powerKw: { type: number }
        weightKg: { type: number }
        color: { type: string }
        importCountryCode: { type: string }
        importDeclarationNumber: { type: string }
        importDate:
          type: string
          format: date
        customsValue: { type: number }
        originCertificateRef: { type: string }
        homologationNumber: { type: string }
        homologationValidUntil:
          type: string
          format: date
    UpdateVehicleRequest:
      type: object
      properties:
        color: { type: string }
        status:
          type: string
          enum: [IMPORTED, IN_STOCK, RESERVED, SOLD, REGISTERED, STOLEN, SCRAPPED, EXPORTED]
        currentWarehouseId:
          type: string
          format: uuid
        homologationNumber: { type: string }
        homologationValidUntil:
          type: string
          format: date
    VehicleOwnership:
      type: object
      properties:
        id: { type: string }
        vehicleId: { type: string }
        clientId: { type: string }
        clientName: { type: string }
        actorId: { type: string }
        actorName: { type: string }
        transferType: { type: string }
        ownershipStartedAt:
          type: string
          format: date-time
        ownershipEndedAt:
          type: string
          format: date-time
        isCurrent: { type: boolean }
        certificateNumber: { type: string }
    VehicleBlacklistEntry:
      type: object
      properties:
        id: { type: string }
        vehicleId: { type: string }
        vin: { type: string }
        reason:
          type: string
          enum: [STOLEN, FRAUD, COUNTERFEIT, NON_COMPLIANT, ADMIN_SEIZURE, INSURANCE_TOTAL_LOSS]
        reasonDetails: { type: string }
        source: { type: string }
        reportedBy: { type: string }
        reportedAt:
          type: string
          format: date-time
        isActive: { type: boolean }
    CreateBlacklistEntryRequest:
      type: object
      required: [vin, reason]
      properties:
        vehicleId:
          type: string
          format: uuid
        vin:
          type: string
        reason:
          type: string
          enum: [STOLEN, FRAUD, COUNTERFEIT, NON_COMPLIANT, ADMIN_SEIZURE, INSURANCE_TOTAL_LOSS]
        reasonDetails: { type: string }
        sourceReference: { type: string }

    # -------------------------------------------------------------------------
    # Stocks
    # -------------------------------------------------------------------------
    StockMovement:
      type: object
      properties:
        id: { type: string }
        vehicleId: { type: string }
        vehicleVin: { type: string }
        actorId: { type: string }
        warehouseId: { type: string }
        warehouseName: { type: string }
        movementType:
          type: string
          enum: [IMPORT, ASSEMBLY, TRANSFER, SALE, RETURN, ADJUSTMENT, SCRAP, RESERVATION, RELEASE]
        quantity: { type: integer }
        referenceDocument: { type: string }
        unitCost: { type: number }
        currencyCode: { type: string }
        notes: { type: string }
        performedBy: { type: string }
        movementDate:
          type: string
          format: date-time
    CreateStockMovementRequest:
      type: object
      required: [vehicleId, movementType]
      properties:
        vehicleId:
          type: string
          format: uuid
        movementType:
          type: string
          enum: [IMPORT, ASSEMBLY, TRANSFER, SALE, RETURN, ADJUSTMENT, SCRAP, RESERVATION, RELEASE]
        warehouseId:
          type: string
          format: uuid
        referenceDocument: { type: string }
        referenceSaleId:
          type: string
          format: uuid
        referenceTransferFrom:
          type: string
          format: uuid
        referenceTransferTo:
          type: string
          format: uuid
        unitCost: { type: number }
        currencyCode: { type: string }
        notes: { type: string }
    InventoryCount:
      type: object
      properties:
        id: { type: string }
        warehouseId: { type: string }
        warehouseName: { type: string }
        status:
          type: string
          enum: [PLANNED, IN_PROGRESS, COMPLETED, RECONCILED, DISCREPANCY_FOUND]
        plannedDate:
          type: string
          format: date
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        totalUnitsExpected: { type: integer }
        totalUnitsCounted: { type: integer }
        discrepancyUnits: { type: integer }
        discrepancyValue: { type: number }
        performedBy: { type: string }

    # -------------------------------------------------------------------------
    # Clients
    # -------------------------------------------------------------------------
...